refactor: swagger picker into common
