+++ src/web/src/views/workspace/WSEditorClientConfig.tsx
@@ -22,7 +22,7 @@ import {
 import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
 import DoDisturbOnRoundedIcon from "@mui/icons-material/DoDisturbOnRounded";
 import AddCircleRoundedIcon from "@mui/icons-material/AddCircleRounded";
-import SwaggerItemSelector from "./SwaggerItemSelector";
+import SwaggerItemSelector from "./common/SwaggerItemSelector";
 import AddRoundedIcon from "@mui/icons-material/AddRounded";
 import type { Plane, Resource } from "./interfaces";
