+++ src/web/src/views/workspace/common/SwaggerItemSelector.tsx
